module-4/activity/poetry/parallel/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-google-genai (>=2.1.12,<3.0.0)",
    "langgraph (>=0.6.8,<0.7.0)",
    "tavily-python (>=0.7.12,<0.8.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "langchain-community (==0.3.30)",
    "langgraph-sdk (>=0.2.9,<0.3.0)",
    "langchain-tavily (>=0.2.11,<0.3.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
